Add config option
