+++ lib/default-config.json
@@ -1,0 +1,3 @@
+{
+  "updateComment": true
+}
