+++ lib/index.js
@@ -2,6 +2,9 @@ const Travis = require('./providers/Travis')
 const Circle = require('./providers/Circle')
 const fs = require('fs')
 const path = require('path')
+const defaultConfig = require('./default-config.json')
+const newComment = require('./new-comment')
+const updateComment = require('./update-comment')
 
 // We already have an instance of handlebars from Probot's hbs dependency
 const { handlebars } = require('hbs')
@@ -11,6 +14,8 @@ module.exports = robot => {
   robot.log('App has started!')
 
   robot.on('status', async context => {
+    const config = await context.config('ci-reporter.yml', defaultConfig)
+
     // Only trigger on failed statuses
     if (context.payload.state === 'failure') {
       let serializer
@@ -31,22 +36,21 @@ module.exports = robot => {
       if (serialized) {
         const { number, data } = serialized
 
-        // Determine if there is already a comment on this PR from ci-reporter
-        const comments = await context.github.issues.getComments(context.issue({ number }))
-        const comment = comments.data.find(comment => comment.user.login === process.env.APP_NAME + '[bot]')
+        const opts = { context, template, data, sha, number }
+        if (config.updateComment) {
+          // Determine if there is already a comment on this PR from ci-reporter
+          const comments = await context.github.issues.getComments(context.issue({ number }))
+          const comment = comments.data.find(comment => comment.user.login === process.env.APP_NAME + '[bot]')
 
-        // If there is, edit that one
-        if (comment) {
-          const lastCommit = /<!--LAST_COMMIT=(.+)-->/g.exec(comment.body)[1]
-          const lastLog = /<!--START_LOG-->([\s\S]+)<!--END_LOG-->/g.exec(comment.body)[1]
-          const oldLogs = /<!--START_OLD_LOGS-->([\s\S]+)<!--END_OLD_LOGS-->/g.exec(comment.body)[1]
-          const body = template({...data, commit: sha, oldLogs, lastLog, lastCommit: lastCommit.substring(0, 7)})
-          return context.github.issues.editComment(context.repo({ number, body, id: comment.id }))
+          // If there is, edit that one
+          if (comment) {
+            opts.comment = comment
+            return updateComment(opts)
+          } else {
+            return newComment(opts)
+          }
         } else {
-          // If there is not, create one
-          const body = template({...data, commit: sha})
-          const issue = { number, body }
-          return context.github.issues.createComment(context.repo(issue))
+          return newComment(opts)
         }
       }
     }
